Changes to Nav, Contact, About, Timeline
Nav:
Added 2 new icons in left of nav bar that link to GitHub and LinkedIn.

Contact:
Started creating a new contact form instead of listing email. This is
not implimented yet but the code is there just commented out.
Contact.php has a new form section added.
main.js has 2 new functions that will assisst the form.
New file mail.php that will be the backup php function for submitting
the form.

About:
Added photoship and illustrator to list of known softwares.

Timeline:
Adjusted the top margin on the SVG so that when on a small screen there
is no margin.