Complete Revamp of Contact Page
The contact page has been redone. There is now a contact form for
submitting emails directly to me. There is also a side bar containing
information that was previously listed.

New icons added to left side of the nav bar.

New Media Queries for tentacle2 so that it doesnt overlap with the now
sider contact section.

Some rewarding in the about section.