Bug Fixes and New SVG
Fixed the navbar overlapping on small screens. I did this by adding a media query that reduces the size of the icons in the navbar and the margin between them.

Temp fix for the landing image appearing too small on large screens. Added media query that increases the dimensions of the image.

Fixed the landing image going off screen on certain aspect ratios. Added JS that checks if the image is off the screen when the landing page is loaded in. If it is it shifts the landing section up to correct for the overlap.

Created a new timeline format for small devices. This is so that timeline text is readable.