Loading, Pic of Me, Fixes
Added a spinning loading symbol that briefly appears when the page is loaded into then fades out. This is accompanied with some JS that forces the screen positioning to the top of the page on load.

I added a picture of me to the about section. Also added a media quiery that shrinks the image when on small screens.

Fixed issue with titles being overlapping with above section on very small screen sizes by adding media queries that increased the padding at the top of the sections.
Added a media query that increased the padding at the bottom of the about section because list appeared too close to base of section.